requirements.txt
streamlit==1.30.0
python-dotenv
google-generativeai
requests